fix: move item methods into their own service
