+++ src/main/java/com/interiordesignplanner/room/RoomService.java
@@ -16,6 +16,7 @@ import com.interiordesignplanner.designer.Designer;
 import com.interiordesignplanner.designer.DesignerService;
 import com.interiordesignplanner.exceptions.ProjectNotFoundException;
 import com.interiordesignplanner.exceptions.RoomNotFoundException;
+import com.interiordesignplanner.inventory.Item;
 import com.interiordesignplanner.mapper.RoomMapper;
 import com.interiordesignplanner.project.Project;
 import com.interiordesignplanner.project.ProjectService;
@@ -251,79 +252,6 @@ public class RoomService {
         }
 
         return roomMapper.toDto(roomRepository.save(existingRoom));
-    }
-
-    /**
-     * Adds item to inventory
-     * 
-     * 
-     * 
-     * @param item   the item is added to the shopping list
-     * @param roomId retrieves the room object to be updated
-     * @throws RoomNotFoundException if the room is not found
-     * @return the updated room, with a new item on the list
-     */
-    @Transactional
-    @PreAuthorize("hasRole('DESIGNER')")
-    public RoomDTO addItem(Long roomId, Item item, String username) {
-
-        Room existingRoom = findRoom(roomId);
-
-        findRoomByDesigner(existingRoom, username);
-
-        item.setOrdered(false);
-
-        existingRoom.getInventory().add(item);
-
-        return roomMapper.toDto(roomRepository.save(existingRoom));
-    }
-
-    /**
-     * Updates item to inventory
-     * 
-     * 
-     * 
-     * 
-     * @param index      retrieves the item to be updated
-     * @param roomId     retrieves the room object to be updated
-     * @param updateItem the updated item
-     * @throws RoomNotFoundException if the room is not found
-     * @return the updated item is added to the inventory
-     */
-    @Transactional
-    @PreAuthorize("hasRole('DESIGNER')")
-    public RoomDTO editItem(Long roomId, Item updateItem, int index, String username) {
-
-        Room existingRoom = findRoom(roomId);
-
-        findRoomByDesigner(existingRoom, username);
-
-        existingRoom.getInventory().set(index, updateItem);
-
-        return roomMapper.toDto(roomRepository.save(existingRoom));
-    }
-
-    /**
-     * Deletes Item from inventory
-     * 
-     * 
-     * 
-     * @param index  using the task's index to delete
-     * @param roomId retrieves the room object to be updated
-     * @throws RoomNotFoundException if the room is not found
-     * @return the updated room
-     */
-    @Transactional
-    @PreAuthorize("hasRole('DESIGNER')")
-    public void deleteItem(Long roomId, int index, String username) {
-
-        Room existingRoom = findRoom(roomId);
-
-        findRoomByDesigner(existingRoom, username);
-
-        existingRoom.getInventory().remove(index);
-
-        roomMapper.toDto(roomRepository.save(existingRoom));
     }
 
     /**
